pyproject.toml
[tool.poetry]
name = "recipe-service"
version = "0.1.0"
description = ""
authors = ["temirovazat <azatbekt5@gmail.com>"]
readme = "README.md"
packages = [{include = "recipe_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.95.1"}
uvicorn = "^0.21.1"
psycopg2-binary = "^2.9.6"
loguru = "^0.7.0"
python-dotenv = "^1.0.0"
black = "^23.3.0"
isort = "^5.12.0"
bandit = "^1.7.5"
flake8 = "^6.0.0"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
asyncpg = "^0.27.0"
passlib = "^1.7.4"
cryptography = "^40.0.2"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-multipart = "^0.0.6"
argon2-cffi = "^21.3.0"
pytest = "^7.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
